add proxy wrapper that sets a default rate
